chore: refactor folder structure and imports
