+++ apps/website/src/components/modules/Home/Sections/Carousel.tsx
@@ -6,9 +6,14 @@ import {
   CarouselNext,
   CarouselPrevious,
 } from '@devlaunchers/components/components/Carousel/Carousel';
-import { Quote } from 'lucide-react';
 
 const testimonials = [
+  {
+    content:
+      "Dev Launchers is a great organization to volunteer with, and I would recommend it to anyone looking for an opportunity to learn new things, grow your skills and experience in Tech and work with great people all around the world. I've been volunteering for 9 months now and it has been a great experience. Members are very collaborative, kind and polite. Whoever wants to learn new stuff, join teams full of great people and work on open source projects that has real impact this is the place for that! One of the best decisions I've ever made.",
+    author: 'John D.',
+    role: 'Team Member',
+  },
   {
     content:
       "Dev Launchers is a great organization to volunteer with, and I would recommend it to anyone looking for an opportunity to learn new things, grow your skills and experience in Tech and work with great people all around the world. I've been volunteering for 9 months now and it has been a great experience. Members are very collaborative, kind and polite. Whoever wants to learn new stuff, join teams full of great people and work on open source projects that has real impact this is the place for that! One of the best decisions I've ever made.",
@@ -18,8 +23,8 @@ const testimonials = [
   {
     content:
       "Dev Launchers is a great organization to volunteer with, and I would recommend it to anyone looking for an opportunity to learn new things, grow your skills and experience in Tech and work with great people all around the world. I've been volunteering for 9 months now and it has been a great experience. Members are very collaborative, kind and polite. Whoever wants to learn new stuff, join teams full of great people and work on open source projects that has real impact this is the place for that! One of the best decisions I've ever made.",
-    author: 'John D.',
-    role: 'Team Member',
+    author: 'Sarah K.',
+    role: 'Project Lead',
   },
   {
     content:
@@ -51,9 +56,19 @@ const TestimonialCarousel = () => {
                 key={index}
                 className="pl-2 md:pl-4 md:basis-2/3 lg:basis-1/2"
               >
-                <div className="bg-[#36525A50] rounded-3xl p-12 h-full border-4 border-[#9ACFDF]">
+                <div
+                  className={`${
+                    index % 2 === 1
+                      ? 'bg-[#36525A50] border-[#9ACFDF]'
+                      : 'bg-[#FF573350] border-[#FFA500]'
+                  } rounded-3xl p-12 h-full border-4`}
+                >
                   <div className="flex flex-col h-full">
-                    <p className="text-[#9ACFDF] mb-8 text-sm flex-grow leading-relaxed">
+                    <p
+                      className={`${
+                        index % 2 === 1 ? 'text-[#9ACFDF]' : 'text-[#FFA500]'
+                      } mb-8 text-sm flex-grow leading-relaxed`}
+                    >
                       {testimonial.content}
                     </p>
                     <div className="flex items-center justify-center gap-3">
@@ -74,6 +89,7 @@ const TestimonialCarousel = () => {
               </CarouselItem>
             ))}
           </CarouselContent>
+
           <div className="mt-12 flex gap-2 z-50 items-center justify-center">
             <CarouselPrevious className="relative flex items-center justify-center left-0 text-white border-white hover:bg-gray-800" />
             <CarouselNext className="relative flex items-center justify-center right-0 text-white border-white hover:bg-gray-800" />
