+++ apps/website/src/components/modules/Home/Sections/Section.tsx
@@ -1,0 +1,7 @@
+import { styles } from '../styles';
+
+export const Section = ({ children, className = '' }) => (
+  <section className={`${styles.section.container} ${className}`}>
+    {children}
+  </section>
+);
