+++ apps/website/src/components/modules/Home/Sections/index.ts
@@ -1,0 +1,7 @@
+export { default as Hero } from './Hero';
+export { default as Opportunities } from './Opportunities';
+export { default as JoinUs } from './JoinUs';
+export { default as TechStack } from './TechStack';
+export { default as Community } from './Community';
+export { default as Features } from './Features';
+export { default as Donate } from './Donate';
