+++ apps/website/src/components/modules/Home/Sections/sections.js
@@ -1,0 +1,36 @@
+export const sections = {
+  hero: {
+    title: 'Collaborate, Build, and Thrive in Tech',
+    description:
+      "Join our global community of 200+ members who've launched their careers to the next level through collaborative, real-world projects.",
+    label: 'Partners and Sponsors',
+    buttonText: 'Explore Opportunities',
+  },
+  opportunities: {
+    title: 'Opportunities for Everyone',
+    description:
+      "Whether you're starting your journey, leading teams, or seeking to give back, we provide the hands-on experience, mentorship, and community you need to succeed.",
+  },
+  join: {
+    title: 'Join Us',
+    description:
+      "Create your profile, discover opportunities, and join our community—all for free! Whether you're just starting out, looking to switch careers, or seeking new challenges, our community is here to support you every step of the way.",
+    buttonText: 'Join Today',
+  },
+  features: {
+    title: 'Why Dev Launchers?',
+    description:
+      "Transform your tech career through real-world experience, collaborative projects, and supportive community. We provide the hands-on opportunities and resources you need to succeed in today's tech industry.",
+  },
+  build: {
+    title: 'We Build with the Best',
+    description:
+      'We collaborate and build using the best tools and software including Figma, React, Next.js and more to launch our members to the next level. Learn how we leverage these technologies in real-world projects and develop competitive skills that keep you at the forefront of the industry.',
+  },
+  donate: {
+    title: 'Transform Tech Today',
+    description:
+      "Together, we're building pathways into technology careers for everyone. Your tax-deductible donation helps create an inclusive community where aspiring developers, designers, product managers, researchers, and QA professionals can turn their potential into impact.",
+    buttonText: 'Donate Today',
+  },
+};
